moved counter to a separate component, added a snackbar notification and deleted an unused theme file
